+++ src/config/migrator.zig
@@ -2,29 +2,29 @@ const std = @import("std");
 const ini = @import("zigini");
 const Save = @import("Save.zig");
 
-const Allocator = std.mem.Allocator;
+pub fn tryMigrateSaveFile(user_buf: *[32]u8, path: []const u8) Save {
+    var save = Save{};
 
-pub fn tryMigrateSaveFile(allocator: Allocator, path: []const u8) Save {
-    var file = std.fs.openFileAbsolute(path, .{ .mode = .read_write }) catch return .{};
+    var file = std.fs.openFileAbsolute(path, .{}) catch return save;
     defer file.close();
 
     const reader = file.reader();
-    const user_length = reader.readIntLittle(u64) catch return .{};
 
-    const user_buffer = allocator.alloc(u8, user_length) catch return .{};
-    defer allocator.free(user_buffer);
+    var user_fbs = std.io.fixedBufferStream(user_buf);
+    reader.streamUntilDelimiter(user_fbs.writer(), '\n', 32) catch return save;
+    const user = user_fbs.getWritten();
+    if (user.len > 0) save.user = user;
 
-    const read_user_length = reader.read(user_buffer) catch return .{};
-    if (read_user_length != user_length) return .{};
+    var session_buf: [20]u8 = undefined;
+    var session_fbs = std.io.fixedBufferStream(&session_buf);
+    reader.streamUntilDelimiter(session_fbs.writer(), '\n', 20) catch {};
 
-    const session_index = reader.readIntLittle(u64) catch return .{};
-
-    const save = .{
-        .user = user_buffer,
-        .session_index = session_index,
-    };
-
-    ini.writeFromStruct(save, file.writer(), null) catch return save;
+    const session_index_str = session_fbs.getWritten();
+    var session_index: ?u64 = null;
+    if (session_index_str.len > 0) {
+        session_index = std.fmt.parseUnsigned(u64, session_index_str, 10) catch return save;
+    }
+    save.session_index = session_index;
 
     return save;
 }