+++ src/tui/components/Desktop.zig
@@ -9,7 +9,6 @@ const Allocator = std.mem.Allocator;
 const EnvironmentList = std.ArrayList(Environment);
 
 const DisplayServer = enums.DisplayServer;
-const ViMode = enums.ViMode;
 
 const termbox = interop.termbox;
 
@@ -17,7 +16,9 @@ const Desktop = @This();
 
 pub const Environment = struct {
     entry_ini: ?Ini(Entry) = null,
-    name: []const u8 = "",
+    name: [:0]const u8 = "",
+    xdg_session_desktop: [:0]const u8 = "",
+    xdg_desktop_names: ?[:0]const u8 = "",
     cmd: []const u8 = "",
     specifier: []const u8 = "",
     display_server: DisplayServer = .wayland,
@@ -25,11 +26,11 @@ pub const Environment = struct {
 
 const DesktopEntry = struct {
     Exec: []const u8 = "",
-    Name: []const u8 = "",
-    DesktopNames: []const u8 = "",
+    Name: [:0]const u8 = "",
+    DesktopNames: ?[]const u8 = null,
 };
 
-pub const Entry = struct { Desktop_Entry: DesktopEntry = DesktopEntry{} };
+pub const Entry = struct { @"Desktop Entry": DesktopEntry = DesktopEntry{} };
 
 allocator: Allocator,
 buffer: *TerminalBuffer,
@@ -56,6 +57,8 @@ pub fn init(allocator: Allocator, buffer: *TerminalBuffer, max_length: u64, lang
 pub fn deinit(self: Desktop) void {
     for (self.environments.items) |*environment| {
         if (environment.entry_ini) |*entry_ini| entry_ini.deinit();
+        if (environment.xdg_desktop_names) |desktop_name| self.allocator.free(desktop_name);
+        self.allocator.free(environment.xdg_session_desktop);
     }
 
     self.environments.deinit();
@@ -67,11 +70,29 @@ pub fn position(self: *Desktop, x: u64, y: u64, visible_length: u64) void {
     self.visible_length = visible_length;
 }
 
-pub fn addEnvironment(self: *Desktop, name: []const u8, cmd: []const u8, display_server: DisplayServer) !void {
+pub fn addEnvironment(self: *Desktop, entry: DesktopEntry, xdg_session_desktop: []const u8, display_server: DisplayServer) !void {
+    var xdg_desktop_names: ?[:0]const u8 = null;
+    if (entry.DesktopNames) |desktop_names| {
+        const desktop_names_z = try self.allocator.dupeZ(u8, desktop_names);
+        for (desktop_names_z) |*c| {
+            if (c.* == ';') c.* = ':';
+        }
+        xdg_desktop_names = desktop_names_z;
+    }
+
+    errdefer {
+        if (xdg_desktop_names) |desktop_names| self.allocator.free(desktop_names);
+    }
+
+    const session_desktop = try self.allocator.dupeZ(u8, xdg_session_desktop);
+    errdefer self.allocator.free(session_desktop);
+
     try self.environments.append(.{
         .entry_ini = null,
-        .name = name,
-        .cmd = cmd,
+        .name = entry.Name,
+        .xdg_session_desktop = session_desktop,
+        .xdg_desktop_names = xdg_desktop_names,
+        .cmd = entry.Exec,
         .specifier = switch (display_server) {
             .wayland => self.lang.wayland,
             .x11 => self.lang.x11,
@@ -83,11 +104,30 @@ pub fn addEnvironment(self: *Desktop, name: []const u8, cmd: []const u8, display
     self.current = self.environments.items.len - 1;
 }
 
-pub fn addEnvironmentWithIni(self: *Desktop, entry_ini: Ini(Entry), name: []const u8, cmd: []const u8, display_server: DisplayServer) !void {
+pub fn addEnvironmentWithIni(self: *Desktop, entry_ini: Ini(Entry), xdg_session_desktop: []const u8, display_server: DisplayServer) !void {
+    const entry = entry_ini.data.@"Desktop Entry";
+    var xdg_desktop_names: ?[:0]const u8 = null;
+    if (entry.DesktopNames) |desktop_names| {
+        const desktop_names_z = try self.allocator.dupeZ(u8, desktop_names);
+        for (desktop_names_z) |*c| {
+            if (c.* == ';') c.* = ':';
+        }
+        xdg_desktop_names = desktop_names_z;
+    }
+
+    errdefer {
+        if (xdg_desktop_names) |desktop_names| self.allocator.free(desktop_names);
+    }
+
+    const session_desktop = try self.allocator.dupeZ(u8, xdg_session_desktop);
+    errdefer self.allocator.free(session_desktop);
+
     try self.environments.append(.{
         .entry_ini = entry_ini,
-        .name = name,
-        .cmd = cmd,
+        .name = entry.Name,
+        .xdg_session_desktop = session_desktop,
+        .xdg_desktop_names = xdg_desktop_names,
+        .cmd = entry.Exec,
         .specifier = switch (display_server) {
             .wayland => self.lang.wayland,
             .x11 => self.lang.x11,
@@ -100,7 +140,7 @@ pub fn addEnvironmentWithIni(self: *Desktop, entry_ini: Ini(Entry), name: []cons
 }
 
 pub fn crawl(self: *Desktop, path: []const u8, display_server: DisplayServer) !void {
-    var iterable_directory = std.fs.openIterableDirAbsolute(path, .{}) catch return;
+    var iterable_directory = std.fs.openDirAbsolute(path, .{ .iterate = true }) catch return;
     defer iterable_directory.close();
 
     var iterator = iterable_directory.iterate();
@@ -110,9 +150,19 @@ pub fn crawl(self: *Desktop, path: []const u8, display_server: DisplayServer) !v
         const entry_path = try std.fmt.allocPrint(self.allocator, "{s}/{s}", .{ path, item.name });
         defer self.allocator.free(entry_path);
         var entry_ini = Ini(Entry).init(self.allocator);
-        var entry = try entry_ini.readToStruct(entry_path);
+        _ = try entry_ini.readFileToStruct(entry_path);
+        errdefer entry_ini.deinit();
 
-        try self.addEnvironmentWithIni(entry_ini, entry.Desktop_Entry.Name, entry.Desktop_Entry.Exec, display_server);
+        var xdg_session_desktop: []const u8 = undefined;
+        const maybe_desktop_names = entry_ini.data.@"Desktop Entry".DesktopNames;
+        if (maybe_desktop_names) |desktop_names| {
+            xdg_session_desktop = std.mem.sliceTo(desktop_names, ';');
+        } else {
+            // if DesktopNames is empty, we'll take the name of the session file
+            xdg_session_desktop = std.fs.path.stem(item.name);
+        }
+
+        try self.addEnvironmentWithIni(entry_ini, xdg_session_desktop, display_server);
     }
 }